Refactoring components
- Break details components out into separate files
- Move into new "details" dir
